server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiosmtplib==2.0.2
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
fqdn==1.5.1
frozenlist==1.4.0
idna==3.4
isoduration==20.11.0
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.4.4
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
python-dateutil==2.8.2
PyYAML==6.0.1
referencing==0.30.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rpds-py==0.10.3
six==1.16.0
uri-template==1.3.0
webcolors==1.13
yarl==1.9.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.113.0
google-auth-httplib2==0.2.0
